+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated `/users/{id}/tenants` endpoint to return all tenants if user is admin
 - Updated creation of user keys only when `identity.key` config value is `true`
+- Updated `TenantUsers` controller `delete` method to allow self to remove from tenant
 
 ### Fixed